[cmp] Build the list of replacement
The list contains a pointer to the instruction to be replaced and a
list of basic block semantically equivalent for each record.

When this list is empty there is no instruction that need to be
replaced with its pool of alternatives. In this case the pass doesn't
modify the function.