[cmp] Add InstructionAlternativeUtils
This file contains the declaration of the InstructionTy that model the
set of instruction that defines alternatives and a set of utility.

CMP_SET_ALTERNATIVE_NUMBER is a macro used to declare the number of
alternatives available for each instruction.

InitializeAlternativesVector is a TMP loop that helps in the build of the
alternative vector.